帝可得后端-商品管理（优化版）：商品管理页面，设备管理页面的货道与商品进行绑定实现

+++ dkd-manage/src/main/java/com/dkd/manage/service/impl/SkuServiceImpl.java
@@ -1,12 +1,17 @@
 package com.dkd.manage.service.impl;
 
 import java.util.List;
+
+import com.dkd.common.exception.ServiceException;
 import com.dkd.common.utils.DateUtils;
+import com.dkd.common.utils.poi.ExcelUtil;
+import com.dkd.manage.service.IChannelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.dkd.manage.mapper.SkuMapper;
 import com.dkd.manage.domain.Sku;
 import com.dkd.manage.service.ISkuService;
+import org.springframework.web.multipart.MultipartFile;
 
 /**
  * 商品管理Service业务层处理
@@ -19,6 +24,9 @@ public class SkuServiceImpl implements ISkuService
 {
     @Autowired
     private SkuMapper skuMapper;
+    @Autowired
+    private IChannelService channelService;
+
 
     /**
      * 查询商品管理
@@ -79,6 +87,12 @@ public class SkuServiceImpl implements ISkuService
     @Override
     public int deleteSkuBySkuIds(Long[] skuIds)
     {
+        //1. 判断商品的id集合是否有关联货道
+        int count = channelService.countChannelBySkuIds(skuIds);
+        if(count>0){
+            throw new ServiceException("此商品被货道关联，无法删除");
+        }
+        //2. 没有关联货道才能删除
         return skuMapper.deleteSkuBySkuIds(skuIds);
     }
 
@@ -92,5 +106,17 @@ public class SkuServiceImpl implements ISkuService
     public int deleteSkuBySkuId(Long skuId)
     {
         return skuMapper.deleteSkuBySkuId(skuId);
+    }
+
+    /**
+     * 批量新增商品管理
+     * @param file
+     * @return 结果
+     */
+    @Override
+    public int insertSkus(MultipartFile file) throws Exception  {
+        ExcelUtil<Sku> util = new ExcelUtil<Sku>(Sku.class);
+        List<Sku> skuList = util.importEasyExcel(file.getInputStream());
+        return skuMapper.insertSkus(skuList);
     }
 }
